Fix issues with nested programs & objects
When updating the properties of the channels arrays, it was always
updating them on *all* the arrays, not just the one you thought you were
targeting.

Make a copy of the array once we start changing its properties, so that
doesn't also update the original channels.

This was causing bugs in setting all previous/next links, even if the
channel isn't 'available', and not picking the correct program if a
channel has 1 program that matches and 1 that doesn't.