Change program blocks when the hour changes!
This uses the `session.currentHour` prop in addition to the program
reset code to change the visible programming across the site.

This also adds in a fallback in the case that the new hour (or any hour)
doesn't have a programBlock.

This doesn't yet reset availablePrograms() so that they're accurately
reflected in the previous/next buttons, but I hope that by refactoring
the way the data is formatted, I can let that fix it instead of doing it
this hacky way.