Use hover to switch eye icon instead of event listeners (mobile fix)
For some reason `onMouseEnter` was making it so that you'd have to click
twice to go to the TV Guide link. So was just using this show/hide SVG
`:hover` approach without setting a media query.

Using the new-to-me `@media (hover: hover)` to only apply the styles on
devices that support hover. Pretty amazing and supported by mobile!