Correctly handle channels with multiple programs
Expect that there may be additional bugs here.

But ultimately, while we were correctly *finding* the featured, active programs, we weren't then saving what we found to the channel data.  It was then continuing to include all programs for a channel in the channel list, regardless of featured or active date.