Pull out Tooltip into its own reusable component
Phew! Instead of copying that HTML structure all over the place, this
lets us pass in the info, and standardize the styles.

The one caveat here is that the `react-overlay` positioning code gets a
little wacky once what we're passing into `<Overlay>` isn't a standard
`div`. That's why `<Tooltip>` has that additional positioning logic for
`positionLeft` and `positionLeft` that get passed in by `<Overlay>`, but
weren't getting applied until manually adding the CSS.

The `ignorePositioning` flag is another workaround, that can tell the
tooltip to ignore the `<Overlay>` positioning code that gets passed in.