Phew. Switch from using old react-overlays to Bootstrap's current Overlays
This is a real pain! But the `react-overlays` library was really out of
date, so we can use `react-bootstrap`s latest version.

Depending on the placement, we can use `<Overlay>` by itself, or a
combination of `<OverlayTrigger>` and `<Popover>` for simpler handling.

Only `<Overlay>` gives us an API to manually close the tooltips, so I
ended up removing the close button across the tooltips. It's possible to
get `<Overlay>`s working for all scenarios, but it was taking a ton of
time and didn't seem worth it, so went with this method for sanity's
sake.

This almost gets the app up to date. Still need to update `react-player`
and fix that fullscreen bug that's happening on mobile.