Move controls into separate container from video so text can lay on top on mobile
When using `transform`, it messes with the `z-index` and the stacking
context of the element and its children. In order to show the program
name text on top of the video (which, we want for now, at least), we
need to have the video and the previous/next/close buttons to be in
separate containers, so the video itself can be "behind" the outside
text, and the buttons can be "in front" of them.