Add separate container for video, after all.
The whole point of adding the second container was to allow the text to
lay over the video, in the case that it's too long. And, it turns out
we don't need the z-index after all, we just need each to be in separate
containers. The 'Info' section still opens correctly on mobile.

Original commit here: https://github.com/singleportrait/locally-grown/commit/fa60a2d68c6ebbff6e19cb8201f314742c1e358a