Use MediaQuery plugin instead of native media query to show/hide content
Even though the content was `display: none`, because the video content
is rendering twice on the page, the mute/unmute button will also mute
and unmute both videos, which causes problems on desktop.

Maybe there will be a way to revive it, by making sure the video is
rendered more conditionally, but for now, rolling that fix back until we
can get a better one out.