WIP: After detecting low battery, try to detect if video is playing
If we detect it's not playing, show the play button and let users click
it to play the video. If we detect it's playing, don't show the button.

This doesn't work consistently and requires a timeout to guess whether
the video is going to successfully load.

Needs some work, but worth pushing up remotely in case some of this code
comes in handy.