Remove custom logic to re-mute between Vimeo videos
Previously, if you came unmuted from a non-Vimeo video to a Vimeo one,
the video wouldn't play. However, when simply removing this code, that
no longer seems to be an issue.

I tested:
- Multiple Vimeo videos in a row, unmuting one and letting them continue
to play
- Vimeo videos split up with Youtube videos; unmuting one and letting
them continue to play worked correctly
- Unmuting a Youtube video on another channel, then coming to this
channel with Vimeo videos playing. Audio worked correctly.

Maybe the `react-player` got an update to play more nicely.

I left the section in in case we need to use it for something else
(especially fixing that "deferred player + URL" video loading issue)