Increase setTimeout for handling rotated screens
On the Android I'm testing, I was consistently seeing that 100ms was too short to successfully have the 'new' phone size. Upping that number to 150ms seemed to make it better.

Also, having the viewport widths and heights within a function meant they weren't getting re-run appropriately. Using the bare numbers instead to determine whether to show the mobile screen or not.