Only show low battery warning on iOS
First, there was a bug on Android where the low battery warning is appearing every time, even when the phone isn't on low battery mode. For some reason, it's not quickly autoplaying the test video—it does eventually, but not for many seconds sometimes. Therefore, that isn't a reliable test for Android.

But secondly, Android doesn't appear to do the same thing as iOS, where it pauses the videos on low battery mode. On a Samsung Galaxy S5, at least, the site's videos continue to play even on low battery mode.

So, this test isn't even necessary for those phones. Therefore, it feels appropriate to only run this test if we detect that they're on an iOS device, iPhone or iPad.