Remove orientationchange listener on unmount; add additional notes on how the mobile recalculation is working
After testing more browsers, we still know that the orientationchange
event timeout isn't 100% consistent, but leaving it like that for now until it's reported again. However, there's a clear path forward on using the resize event if we want; perhaps we could throttle it so it's not ridiculous on desktop.