Bugfix for playing custom timestamps when first video of the hour
Custom timestamps weren't working on videos if they're one of several videos in the hour. This was because we weren't checking, when iterating through videos, whether the video is the first in the hour or not. We should only be running the additional code if it's the first video in the hour.