Fix bug with window viewport height & width on mobile
When going back and forth between the TV guide and channel pages, the
`window.innerHeight` and `window.innerWidth` become much larger than the
viewport size, because of the scrolling effect on the TV Guide.

There's a newer unit, `window.visualViewport.width` & height, that we
can use instead, that will more reliably give us the size to make our
media queries on.

Now that the `Program` component uses hooks, it might be worth a try to
use them for the hacky media query required for the mobile rotating
issue.