Fix bug with viewing TV guide for channel/program with no program blocks
Also fix a bug where you could see the wrong program on the channel page
itself.

Now, check at the very beginning when `configuringChannels` to make sure
that all programs that get passed through are available for today's
date.

Then, we can remove that check specifically for `featured` channels, as
it's now checking for all of them.