Require additional name & email fields on events/${eventsId}/members/${userID}
This is so that we don't have to write crazy permissions in Firestore to
allow event admins to view users that have those event IDs. In the
future we'll still probably want to save that on the users table so we
can show a user all screenings they're registered for, but we'll cross
that bridge when we get to it.