Final visual setup of Stripe payments and displaying on confirmation screen
There's some WIP in here, playing with using Firebase Functions to create a simpler payment intent, in order to use `stripe.confirmCardPayment`, while trying to avoid ever needing to save the payment methods, and simplify the code overall.

It was taking too much work and testing, and meant we'd still need to save everything to Firestore somewhere, so I mostly ditched it, at least for now, but it might be useful later.